requirements/requirements.txt
absl-py>=0.13.0
chex>=0.0.8
dm-haiku>=0.0.4
jax>=0.2.18
jaxlib>=0.1.69
numpy>=1.21.1
opt-einsum>=3.3.0
optax>=0.0.9
six>=1.16.0
toolz>=0.11.1
